WebScraper
Used some java to get as many 1 move fork url's as we want.  What I wrote asks for how many pages you want to download, and it will put all of them in a csv file.  I attached a csv file from 20 pages worth of content (few thousand entries I believe).  I can work on this later in the week to extract the moves in the way that we want.